feat: 축제 리스트 조회 - 캘린더 페이지 Service, Repository 구현 및 테스트

+++ src/main/java/com/swyp10/domain/festival/entity/FestivalBasicInfo.java
@@ -31,9 +31,9 @@ public class FestivalBasicInfo {
     @Column(columnDefinition = "TEXT")
     private String firstimage2;
 
-    private Long mapx;
+    private Double mapx;
 
-    private Long mapy;
+    private Double mapy;
 
     private String modifiedtime;
